backend/pyproject.toml
[project]
name = "simple"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.29",
    "langgraph>=0.6.4",
    "python-dotenv>=1.1.1",
]
